labels_nervosinho/requirements.txt
Pillow==9.3.0
Pillow==10.2.0
qrcode==7.4.2
tqdm==4.65.0
